bh(mcp.project)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


bh(sorted, q, m, adjust = FALSE, m0 = m, pi0, constant = 1)

Arguments

sorted 	 ~~Describe sorted here~~

q 	 ~~Describe q here~~

m 	 ~~Describe m here~~

adjust 	 ~~Describe adjust here~~

m0 	 ~~Describe m0 here~~

pi0 	 ~~Describe pi0 here~~

constant 	 ~~Describe constant here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (sorted, q, m, adjust = F, m0 = m, pi0, constant = 1) 
{
    cat("Calling bh \n")
    flush.console
    if (missing(m0) & !missing(pi0)) 
        m0 = pi0 * m
    else {
        criticals = (1:m) * q/(m0 * constant)
        cutoff = reject(sorted, criticals)
        rejected = sorted <= cutoff$cutoff
        adjusted = rep(NA, m)
        if (adjust) 
            adjusted = bh.adjust(sorted, m = m, m0 = m0, constant = constant)
        multiple.pvals = data.frame(original.pvals = sorted, 
            criticals = criticals, rejected = rejected, adjusted.pvals = adjusted)
        output = list(Cutoff = cutoff, Pvals = multiple.pvals)
        return(output)
    }
  }

[Package mcp.project version 0.8 Index]