fdr(mcp.project)	R Documentation

Multiple hypothesis testing while controlling the False Discovery rate

Description

The main function of package mcp.project. Given a p-value vector, the function will return the rejected hypothesis, the cutoff, and the FDR adjusted p-values according to the procedure selected.

Usage


fdr(x, q, method = "BH", pi0, lambda = 0.5, m0)  


Arguments

x 	 A vector if p-values

q 	 The False Discovery Rate to control for

method 	 The desired procedure. See Details.

pi0 	 ~~Describe pi0 here~~

lambda 	 The threshold for Story's m0 estimator.

m0 	 The number of true null hypothsis. Usually unknown.

Details

The fdr function is the main function in the package mcp.project. Given a p-value vector, it can apply several FDR controlling procedures and return the rejected hypothesis, the cutoff and the adjusted p-values. The methods supplied and chosen by the 'method' parameter are the following:

'BH'

the plain Linear Step-up procedure set as default. Known to control the FDR in the case in independant tests and possitively correlated test statistics.

'General Dependency'

A "worst case" procedure when the dependency structure between the p-values is unknown.

'Oracle'

When m0 is (miraculously) known it can be used to increase the power of the Linear Step Up procedure while still controlling the False Discovery Rate (as long as the dependance structure is such that a linear step up will control the FDR). This has been shown to be the assymptotically most "powerfull" procedure which controls the FDR (Genovese & Wasserman 2002).

'BH Adaptive'

An adaptive procedure which gains power by first estimating m0 using an FWE derived estimator and then plugging it in a Linear Step-up. FDR control under independance has not been proven yet is controlled in simulation studies. See the references for further details.

'ST Adaptive'

An adaptive procedure. First estimating m0 and then and then plugging it in a Linear Step-up. This procedure requires a threshold (lambda) which is set at 0.5 by default. FDR is *not* controlled if all null hypothesis are true. See the references for further details.

'Two Stage'

An adaptive procedure. First estimating m0 using a conservative linear step-up and then using it to increase the power of another linear step-up. In the case of independent test statistics, this procedure has been shown to control the FDR.

'Multiple Step Down'

A non-linear step-down procedure. Enjoys more power then most procedures and like the "Two Stage" procedure, has been proven to control the FDR in the indepenent-test-statistics case.

Value

method 	 The procedure used.

q 	 The FDR level controlled for.

Cutoff 	 A list containing the p-value cutoff and it's index.

Pvals 	 A data frame containing the original p-values, the critical value each has been compared to, an indicator for the rejection of each hypothesis and the procedure specific adjusted p-values.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Jonathan Rosenblatt.

References

Adaptive linear step-up procedures that control the false discovery rate. Y. Benjamini, A.M. Krieger, D. Yekutieli (Biometrika 2006)

See Also

http://www.math.tau.ac.il/~ybenja/fdr/index.htm http://strimmerlab.org/notes/fdr.html

Examples


test=runif(100)^3

#Linear Step up
bh.1=fdr(x=test,q=0.1,'BH')
plot(bh.1)
summary(bh.1)
table(bh.1$Pvals[['adjusted.pvals']]<=bh.1$q)

# General Dependancy structure
general.1=fdr(x=test,q=0.1,method='General Dependency')
plot(general.1)
summary(general.1)

# New Two Stage procedure
twostage.1=fdr(x=test,0.1,'Two Stage')
plot(twostage.1)
summary(twostage.1)

# If m0 is somehow known
oracle.1=fdr(x=test,q=0.1,'Oracle',m0=30)
plot(oracle.1);summary(oracle.1)

#Adaptive procedure with Story's m0 estimator
adaptive.st.1=fdr(x=test,q=0.1,'ST Adaptive')
plot(adaptive.st.1)
summary(adaptive.st.1)

#Adaptive procedure with Benjamini and Hochberg's m0 estimator
adaptive.bh.1=fdr(x=test,0.1,'BH Adaptive')
plot(adaptive.bh.1)
summary(adaptive.bh.1)

# (Non Linear) Step down procedure
multiple.1=fdr(x=test,q=0.1,'Multiple Step Down')
plot(multiple.1)
summary(multiple.1)

[Package mcp.project version 0.8 Index]