Get JAGS code for a prior

Get JAGS code for a prior

get_prior_str(prior, i, varying_group = NULL)

Arguments

prior
Named list. Names are parameter names (cp_i, int_i, xvar_i, `sigma``) and the values are either

A JAGS distribution (e.g., int_1 = "dnorm(0, 1) T(0,)") indicating a conventional prior distribution. Uninformative priors based on data properties are used where priors are not specified. This ensures good parameter estimations, but it is a questionable for hypothesis testing. mcp uses SD (not precision) for dnorm, dt, dlogis, etc. See details. Change points are forced to be ordered through the priors using truncation, except for uniform priors where the lower bound should be greater than the previous change point, dunif(cp_1, MAXX).

A numerical value (e.g., int_1 = -2.1) indicating a fixed value.

A model parameter name (e.g., int_2 = "int_1"), indicating that this parameter is shared - typically between segments. If two varying effects are shared this way, they will need to have the same grouping variable.

i
The index in prior to get code for

varying_group
String or NULL (default). Null indicates a population- level prior. String indicates a varying-effects prior (one for each group level).

Value

A string