RowSums of element-wise products. Like an inner product, just vectorized for many y. This ensures R <–> JAGS code compatibility

RowSums of element-wise products. Like an inner product, just vectorized for many y. This ensures R <–> JAGS code compatibility

Usage

inprod(x, y)

Arguments

x

A matrix

y

A matrix

Value

A vector

Author

Jonas Kristoffer Lindeløv jonas@lindeloev.dk