mcp.project-package(mcp.project)	R Documentation

Applies different False Discovery Rate controlling procedures

Description

The package offers several FDR controlling procedures. Given a p-value vector it will return the rejected hypothsis, the cutoff and the adjusted p-values.

Details

Package: 	 mcp.project

Type: 	 Package

Version: 	 1.0

Date: 	 2008-09-03

License: 	 What license is it under?

LazyLoad: 	 yes

The main function in this package is fdr which test hypothsis using different FDR controlling procedures. The output is an object of class FDR which can be examined with the generic plot and summary functions.

Author(s)

Jonathan Rosenblatt

Maintainer: Jonathan <john.ros@gmail.com>

References

Adaptive linear step-up procedures that control the false discovery rate. Y. Benjamini, A.M. Krieger, D. Yekutieli (Biometrika 2006)

See Also

http://www.math.tau.ac.il/~ybenja/fdr/index.htm http://strimmerlab.org/notes/fdr.html

Examples



#Generating test data
test=runif(100)^3

#Linear Step up
bh.1=fdr(x=test,q=0.1,'BH')
plot(bh.1)
summary(bh.1)
table(bh.1$Pvals[['adjusted.pvals']]<=bh.1$q)

[Package mcp.project version 0.8 Index]