plot.fdr(mcp.project)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


plot(fdr)

Arguments

fdr 	 ~~Describe fdr here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (fdr) 
{
    data = fdr$Pvals
    ord = order(data$original.pvals)
    m = length(ord)
    data = data[ord, ]
    devAskNewPage(ask = T)
    plot(data$original.pvals, cex = 50/m, col = as.numeric(data$rejected) + 
        1, xlab = "Order of P-values", ylab = "", lty = 2, main = "Sorted pvals")
    lines(data$criticals, col = "red")
    abline(h = fdr$Cutoff)
    plot(data$adjusted.pvals ~ data$original.pvals, cex = 50/m, 
        main = "Adjusted p-vals", col = as.numeric(data$rejected) + 
            1, ylim = c(0, 1), xlab = "Original P-values", ylab = "Adjusted P-Values")
    abline(h = fdr$q)
    axis(side = 2, at = fdr$q, labels = "q", las = 2)
    devAskNewPage(ask = F)
  }

[Package mcp.project version 0.8 Index]