Run parallel MCMC sampling using JAGS.

Run parallel MCMC sampling using JAGS.

run_jags(data, jags_code, params, ST, cores, sample = sample,
  model_file = "tmp_jags_code.txt", ...)

Arguments

data
Data.frame or tibble in long format.

jags_code
A string. JAGS model, usually returned by make_jagscode().

params
Character vector of parameters to be sampled.

ST
A segment table (tibble), returned by get_segment_table. Only really used when the model contains varying effects.

cores
Positive integer or "all". Number of cores.

1: serial sampling

>1: parallel sampling on this number of cores. Ideally set chains to the same value.

"all": use all cores but one.

sample
One of TRUE, FALSE, or "prior". Set to FALSE to get an mcpfit object without samples. This is useful if you only want to check priors, the JAGS model, etc. Set to "prior" if you want to do a prior predictive check.

model_file
A temporary file. Makes parallel sampling possible samples that are discarded between n.adapt and n.iter (to improve convergence)..

...
Parameters for jags.parfit which channels them to jags.fit.

Value

mcmc.list

Examples

if (FALSE) {
run_jags(data, model, params)
}