st.m0.estimate(mcp.project)	R Documentation

~~function to do ... ~~

Description

Utility function for the "fdr" function in package mcp.project.

Usage


st.m0.estimate(sorted, m)

Arguments

sorted 	 ~~Describe sorted here~~

m 	 ~~Describe m here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (sorted, m) 
{
    k = 2:m
    m0.k = (m + 1 - k)/(1 - sorted[-1])
    diffs = diff(m0.k, lag = 1)
    indicators = diffs > 0
    optimal.k = min(k[indicators])
    m0 = min(ceiling(m0.k[optimal.k]), m)
    stopifnot(m0 <= m, m0 > 0)
    return(m0)
  }

[Package mcp.project version 0.8 Index]