two.stage.adjust(mcp.project)	R Documentation

~~function to do ... ~~

Description

Utility function for the "fdr" function in package mcp.project.

Usage


two.stage.adjust(sorted, r = 0, patience = 4)

Arguments

sorted 	 ~~Describe sorted here~~

r 	 ~~Describe r here~~

patience 	 ~~Describe patience here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (sorted, r = 0, patience = 4) 
{
    m = length(sorted)
    adjusted = rep(0, m)
    adjusted.q = solve.q(sorted = sorted, m = m, j = 1, r = 0)
    checking = adjusted.q
    if (sum(bh(sorted, adjusted.q[1]/(1 + adjusted.q[1]), m = m)$Pvals[["rejected"]]) == 
        m) {
        adjusted.q = rep(adjusted.q[1], m)
        return(adjusted.q)
    }
    else {
        for (j in 1:m) {
            delta.r = 1
            delta.q = 1
            new.q = adjusted.q[j]
            r.new = sum(bh(sorted, new.q/(1 + new.q), m = m)$Pvals[["rejected"]])
            counter = 0
            max.q = 0
            while (delta.r > 0 & delta.q > 0) {
                old.q = new.q
                r.old = r.new
                new.q = solve.q(sorted = sorted, m = m, j = j, 
                  r = r.old)[j]
                r.new = sum(bh(sorted, new.q/(1 + new.q), m = m)$Pvals[["rejected"]])
                delta.r = abs(r.new - r.old)
                delta.q = abs(new.q - old.q)
                counter = counter + 1
                if (counter > patience & max.q != new.q) 
                  max.q = max(max.q, new.q)
                else if (counter > patience & max.q == new.q) 
                  break
            }
            adjusted.q[j] = min(new.q, 1)
            adjusted.q[min(j + 1, m)] = adjusted.q[j]
            stopifnot(any(adjusted.q[j] <= checking[j]))
        }
        temp.min = adjusted.q[m]
        for (i in m:1) {
            if (adjusted.q[i] <= temp.min) 
                temp.min = adjusted.q[i]
            else adjusted.q[i] = temp.min
        }
        return(adjusted.q)
    }
  }

[Package mcp.project version 0.8 Index]