Fitted and predicted values of mcp models fits

Evaluate the model on data, either summarised (per data-row) or per draw. You can use draws from the prior (prior = TRUE), select the parameter to predict from (“)

Usage

# S3 method for class 'mcpfit'
predict(
  object,
  newdata = NULL,
  summary = TRUE,
  probs = TRUE,
  rate = TRUE,
  prior = FALSE,
  varying = TRUE,
  arma = TRUE,
  ndraws = NULL,
  samples_format = "tidy",
  nsamples = lifecycle::deprecated(),
  ...
)

# S3 method for class 'mcpfit'
fitted(
  object,
  newdata = NULL,
  summary = TRUE,
  probs = TRUE,
  rate = TRUE,
  prior = FALSE,
  dpar = "epred",
  varying = TRUE,
  arma = TRUE,
  ndraws = NULL,
  samples_format = "tidy",
  scale = "response",
  nsamples = lifecycle::deprecated(),
  ...
)

# S3 method for class 'mcpfit'
log_lik(
  object,
  newdata = NULL,
  summary = TRUE,
  probs = TRUE,
  rate = TRUE,
  prior = FALSE,
  varying = TRUE,
  arma = TRUE,
  ndraws = NULL,
  samples_format = "tidy",
  nsamples = lifecycle::deprecated(),
  ...
)

# S3 method for class 'mcpfit'
residuals(
  object,
  newdata = NULL,
  summary = TRUE,
  probs = TRUE,
  prior = FALSE,
  varying = TRUE,
  arma = TRUE,
  ndraws = NULL,
  nsamples = lifecycle::deprecated(),
  ...
)

Arguments

object

An mcpfit object.

newdata

A tibble or a data.frame containing predictors in the model. Weighted Gaussian predictions and log-likelihoods also require the weights column. If NULL (default), the original data is used.

summary

Summarise at each x-value

probs

Vector of quantiles. Only in effect when summary == TRUE.

rate

Boolean. For binomial models, plot on raw data (rate = FALSE) or response divided by number of trials (rate = TRUE). If FALSE, linear interpolation on trial number is used to infer trials at a particular x.

prior

TRUE/FALSE. Plot using prior samples? Useful for mcp(..., sample = "both")

varying

One of:

TRUE All varying effects (fit$pars$varying).

FALSE No varying effects (c()).

"cp" or "predictor": All varying effects belonging to that part of the model.

Character vector: Only include specified varying parameters - see fit$pars$varying.

arma

Whether to include AR and MA effects.

TRUE Compute the GARMA residual recurrence. Requires the response variable in newdata.

FALSE Disregard AR and MA effects. For family = gaussian(), predict() uses only sigma for residuals.

ndraws

Integer or NULL. Number of posterior draws to return/summarise. If there are varying effects, this is the number of draws from each varying group. NULL means "all". Ignored if both are FALSE. More samples trade speed for accuracy.

samples_format

One of "tidy" or "matrix". Controls the output format when summary == FALSE. See more under "value"

nsamples

Deprecated. Use ndraws instead.

...

Currently ignored.

dpar

What distributional parameter to evaluate. This is only relevant when type == "fitted". E.g.,

"epred" (default): Expected value of the full model (or NULL for compatibility with brms etc.).

"mu": The central tendency which is often the mean after applying the link function.

"sigma": The standard deviation of the residuals.

"ar1", "ar2", "ma1", "ma2", etc. depending on which AR or MA coefficient you want to evaluate.

scale

One of

"response": return on the observed scale, i.e., after applying the inverse link function.

"linear": return on the parameter scale (where the linear trends are modelled). A linear scale is only applicable when type == "fitted" and dpar is not NULL.

Value

If summary = TRUE: A tibble with the posterior mean for each row in newdata, If newdata is NULL, the data in fit$data is used.

If summary = FALSE and samples_format = "tidy": A tidybayes tibble with all the posterior draws (Nd) evaluated at each row in newdata (Nn), i.e., with Nd x Nn rows. If there are varying effects, the returned data is expanded with the relevant levels for each row.

The return columns are:

Predictors from newdata.

Draw descriptors: ".chain", ".iteration", ".draw" (see the posterior and tidybayes packages), and data_row, the row number in the evaluated newdata.

Draw values: one column for each parameter in the model.

The estimate. Either ".epred", ".prediction", ".residual", or ".loglik" (matching tidybayes/ggdist conventions).

If summary = FALSE and samples_format = "matrix": An N_draws X nrows(newdata) matrix with fitted/predicted values (depending on type). This format is used by brms and it's useful as yrep in bayesplot::ppc_* functions.

Details

residuals(fit) is equivalent to fit$data[, fit$data$yvar] - fitted(fit, ...) (or newdata[, fit$data$yvar] - fitted(fit, ...)), but with fixed arguments for fitted: rate = FALSE, dpar = 'epred', samples_format = 'tidy'.

Functions

predict(mcpfit): Predictive Distribution

fitted(mcpfit): Expected distribution

log_lik(mcpfit): Pointwise log-likelihood

residuals(mcpfit): Residual distribution

See also

Author

Jonas Kristoffer Lindeløv jonas@lindeloev.dk

Examples

fitted(demo_fit)  # Expected value at each demo_fit$data at response-level
#>            time     fitted     error       Q2.5       Q97.5
#> 1   91.48060435 -3.5747036 0.7559911 -5.0171110 -2.08894192
#> 2   93.70754133 -4.1776285 0.8482945 -5.7852203 -2.54436000
#> 3   28.61395348 11.0198058 1.0897588  9.0408552 13.05466062
#> 4   83.04476261 -1.2907689 0.6569355 -2.5742401  0.01588497
#> 5   64.17455189 25.0238124 0.9117129 23.2659047 26.85500429
#> 6   51.90959491 20.1288619 0.6039014 18.9323044 21.27390910
#> 7   73.65883146  1.2503947 1.0289446 -0.7432435  3.23043810
#> 8   13.46665972  8.9796028 0.8753663  7.2610057 10.62244501
#> 9   65.69922904 25.6323118 0.9753114 23.7580444 27.59086621
#> 10  70.50647840  2.1038683 1.2059656 -0.2792670  4.44676774
#> 11  45.77417762 17.6802137 0.6814187 16.2006300 18.86984631
#> 12  71.91122517  1.7235447 1.1254177 -0.5106579  3.92722175
#> 13  93.46722472 -4.1125648 0.8374898 -5.7080100 -2.49729847
#> 14  25.54288243 10.1554838 0.9286015  8.5146433 12.08742074
#> 15  46.22928225 17.8618462 0.6703099 16.4090220 19.04004453
#> 16  94.00145228 -4.2572025 0.8617510 -5.8810475 -2.59000745
#> 17  97.82264284 -5.2917584 1.0561177 -7.3021375 -3.26939788
#> 18  11.74873617  8.9600279 0.8992556  7.1147403 10.62244501
#> 19  47.49970816 18.3688738 0.6433229 17.0379111 19.52916922
#> 20  56.03327462 21.7746245 0.6500161 20.5221115 23.00770553
#> 21  90.40313873 -3.2829887 0.7187068 -4.6484215 -1.88919190
#> 22  13.87101677  8.9850151 0.8701323  7.2775697 10.62244501
#> 23  98.88917289 -5.5805126 1.1152694 -7.7150078 -3.42765356
#> 24  94.66682326 -4.4373462 0.8931385 -6.1220646 -2.69952002
#> 25   8.24375581  8.9326990 0.9504264  6.8184980 10.62244501
#> 26  51.42117843 19.9339346 0.6037324 18.7184677 21.07939801
#> 27  39.02034671 14.9849390 0.9106297 12.9739794 16.45541071
#> 28  90.57381309 -3.3291974 0.7242407 -4.7029897 -1.92466703
#> 29  44.69696281 17.2502967 0.7104525 15.6815560 18.47398370
#> 30  83.60042600 -1.4412102 0.6482975 -2.6999537 -0.15302789
#> 31  73.75956178  1.2231228 1.0235325 -0.7608466  3.19295067
#> 32  81.10551413 -0.7657334 0.7025357 -2.1348146  0.60159050
#> 33  38.81082828 14.9014294 0.9186842 12.8726923 16.38709684
#> 34  68.51697294 26.7568748 1.0995989 24.6497538 29.01572667
#> 35   0.39483388  8.9217312 0.9789504  6.6601356 10.62244501
#> 36  83.29160803 -1.3576003 0.6528391 -2.6318861 -0.05982645
#> 37   0.73341469  8.9217674 0.9788095  6.6601356 10.62244501
#> 38  20.76589728  9.3264515 0.7747041  7.8189878 10.85898989
#> 39  90.66014078 -3.3525699 0.7270952 -4.7310166 -1.94173823
#> 40  61.17786434 23.8278329 0.7970625 22.2835256 25.39614489
#> 41  37.95592405 14.5607928 0.9517308 12.4352139 16.09085344
#> 42  43.57715850 16.8033823 0.7443177 15.1445977 18.05653402
#> 43   3.74310329  8.9223761 0.9767347  6.6601356 10.62244501
#> 44  97.35399138 -5.1648748 1.0306855 -7.1256890 -3.19991296
#> 45  43.17512489 16.6429304 0.7572986 14.9626342 17.92070508
#> 46  95.75765966 -4.7326812 0.9470826 -6.5370177 -2.91186594
#> 47  88.77549055 -2.8423163 0.6738603 -4.1540726 -1.51845464
#> 48  63.99787695 24.9533013 0.9045444 23.2115638 26.75422147
#> 49  97.09666104 -5.0952048 1.0168804 -7.0289149 -3.16174918
#> 50  61.88382073 24.1095804 0.8226124 22.5197818 25.74211874
#> 51  33.34272113 12.7326610 1.1215631 10.2711171 14.56161185
#> 52  34.67482482 13.2567070 1.0784461 10.8488087 15.01309780
#> 53  39.84854114 15.3152892 0.8782799 13.3743787 16.74570242
#> 54  78.46927757 -0.0519940 0.7964187 -1.6309773  1.49381495
#> 55   3.89364911  8.9224234 0.9765977  6.6601356 10.62244501
#> 56  74.87953862  0.9198983 0.9646297 -0.9779447  2.77749677
#> 57  67.72768302 26.4418688 1.0640290 24.3948787 28.61843382
#> 58  17.12643304  9.0671073 0.8132455  7.5029081 10.64008007
#> 59  26.10879638 10.2960136 0.9599662  8.5984387 12.26531084
#> 60  51.44129347 19.9419625 0.6037163 18.7283032 21.08708779
#> 61  67.56072745 26.3752368 1.0565745 24.3442316 28.52663944
#> 62  98.28171979 -5.4160497 1.0813721 -7.4791333 -3.33766646
#> 63  75.95442676  0.6288812 0.9105813 -1.1695524  2.39027852
#> 64  56.64884241 22.0202979 0.6631426 20.7624499 23.28183526
#> 65  84.96897186 -1.8117327 0.6362753 -3.0313753 -0.54960438
#> 66  18.94739354  9.1671163 0.7811543  7.6751175 10.70750059
#> 67  27.12866147 10.5713994 1.0162683  8.7849613 12.58487090
#> 68  82.81584852 -1.2287923 0.6610972 -2.5169522  0.07972032
#> 69  69.32048204 27.0775557 1.1363272 24.9002709 29.42661141
#> 70  24.05447396  9.8238418 0.8597300  8.2267129 11.62561776
#> 71   4.29887960  8.9225881 0.9760904  6.6601356 10.62244501
#> 72  14.04790941  8.9875754 0.8677659  7.2775697 10.62244501
#> 73  21.63854151  9.4279187 0.7867739  7.9227223 10.99934786
#> 74  47.93985642 18.5445372 0.6354550 17.2319155 19.70586549
#> 75  19.74103423  9.2289563 0.7737628  7.7228191 10.76695065
#> 76  71.93558377  1.7169498 1.1240425 -0.5146494  3.91693426
#> 77   0.78847387  8.9217783 0.9787673  6.6601356 10.62244501
#> 78  37.54899646 14.3987262 0.9676129 12.2334478 15.95392728
#> 79  51.44077083 19.9417539 0.6037167 18.7280477 21.08688800
#> 80   0.15705542  8.9217312 0.9789504  6.6601356 10.62244501
#> 81  58.16040025 22.6235614 0.7011198 21.3024784 23.96958630
#> 82  15.79052082  9.0224660 0.8395113  7.3796430 10.62366310
#> 83  35.90283059 13.7439409 1.0313888 11.4363384 15.40912451
#> 84  64.56318784 25.1789172 0.9276361 23.3986002 27.04055604
#> 85  77.58233626  0.1881381 0.8344184 -1.4536316  1.80950631
#> 86  56.36468416 21.9068902 0.6569044 20.6421242 23.15906489
#> 87  23.37033986  9.6940957 0.8337398  8.1128514 11.42205371
#> 88   8.99805163  8.9370362 0.9407236  6.8788037 10.62244501
#> 89   8.56120649  8.9344361 0.9464939  6.8334543 10.62244501
#> 90  30.52183695 11.6667789 1.1477596  9.4416037 13.65482553
#> 91  66.74265147 26.0487422 1.0204251 24.0905061 28.10821794
#> 92   0.02388966  8.9217312 0.9789504  6.6601356 10.62244501
#> 93  20.85699569  9.3361554 0.7754953  7.8224412 10.86184329
#> 94  93.30341273 -4.0682140 0.8302308 -5.6481602 -2.46287362
#> 95  92.56447486 -3.8681527 0.7986215 -5.3897176 -2.31521572
#> 96  73.40943010  1.3179181 1.0424197 -0.7099665  3.32565470
#> 97  33.30719834 12.7187892 1.1225301 10.2549616 14.54984671
#> 98  51.50633298 19.9679198 0.6036780 18.7556029 21.11195179
#> 99  74.39746463  1.0504158 0.9896860 -0.8810100  2.95777638
#> 100 61.91592400 24.1223929 0.8237979 22.5296189 25.75920344
residuals(demo_fit)  # Residuals at each demo_fit$data at response-level
#>        response        time    residuals     error         Q2.5      Q97.5
#> 1   -3.00841981 91.48060435  0.566283829 0.7559911  -0.91947789  2.0086912
#> 2   -7.87686403 93.70754133 -3.699235533 0.8482945  -5.33250403 -2.0916437
#> 3   16.30291008 28.61395348  5.283104253 1.0897588   3.24824946  7.2620549
#> 4   -0.03735530 83.04476261  1.253413620 0.6569355  -0.05324027  2.5368848
#> 5   27.44631853 64.17455189  2.422506135 0.9117129   0.59131424  4.1804138
#> 6   22.06100045 51.90959491  1.932138593 0.6039014   0.78709135  3.1286961
#> 7    1.98538897 73.65883146  0.734994285 1.0289446  -1.24504913  2.7286325
#> 8   10.35933155 13.46665972  1.379728707 0.8753663  -0.26311346  3.0983259
#> 9   15.87725419 65.69922904 -9.755057631 0.9753114 -11.71361202 -7.8807903
#> 10   1.03823613 70.50647840 -1.065632199 1.2059656  -3.40853161  1.3175031
#> 11  16.41815024 45.77417762 -1.262063464 0.6814187  -2.45169606  0.2175203
#> 12   0.35867723 71.91122517 -1.364867482 1.1254177  -3.56854452  0.8693351
#> 13  -2.36615003 93.46722472  1.746414717 0.8374898   0.13114844  3.3418599
#> 14  15.59894731 25.54288243  5.443463531 0.9286015   3.51152657  7.0843040
#> 15  15.20547289 46.22928225 -2.656373304 0.6703099  -3.83457164 -1.2035491
#> 16   0.40988007 94.00145228  4.667082539 0.8617510   2.99988752  6.2909275
#> 17  -4.22113609 97.82264284  1.070622271 1.0561177  -0.95173821  3.0810014
#> 18  14.15402439 11.74873617  5.193996481 0.8992556   3.53157939  7.0392841
#> 19  22.43276835 47.49970816  4.063894537 0.6433229   2.90359913  5.3948572
#> 20  25.90014996 56.03327462  4.125525449 0.6500161   2.89244443  5.3780384
#> 21  -8.25310350 90.40313873 -4.970114811 0.7187068  -6.36391160 -3.6046820
#> 22   9.63925445 13.87101677  0.654239403 0.8701323  -0.98319055  2.3616848
#> 23  -3.28376193 98.88917289  2.296750664 1.1152694   0.14389163  4.4312458
#> 24  -8.74745808 94.66682326 -4.310111900 0.8931385  -6.04793806 -2.6253935
#> 25   7.82868474  8.24375581 -1.104014271 0.9504264  -2.79376027  1.0101868
#> 26  23.03457521 51.42117843  3.100640619 0.6037324   1.95517720  4.3161075
#> 27  17.58288831 39.02034671  2.597949305 0.9106297   1.12747759  4.6089089
#> 28  -2.25969226 90.57381309  1.069505104 0.7242407  -0.33502524  2.4432974
#> 29  13.80537622 44.69696281 -3.444920512 0.7104525  -4.66860749 -1.8761798
#> 30  -7.11920879 83.60042600 -5.677998586 0.6482975  -6.96618090 -4.4192551
#> 31   5.29891568 73.75956178  4.075792900 1.0235325   2.10596501  6.0597623
#> 32  -1.18941708 81.10551413 -0.423683722 0.7025357  -1.79100757  0.9453976
#> 33  14.75917506 38.81082828 -0.142254301 0.9186842  -1.62792178  1.8864828
#> 34  28.77490032 68.51697294  2.018025483 1.0995989  -0.24082635  4.1251465
#> 35   5.22268442  0.39483388 -3.699046784 0.9789504  -5.39976059 -1.4374512
#> 36  -0.21033401 83.29160803  1.147266271 0.6528391  -0.15050757  2.4215520
#> 37   9.13144062  0.73341469  0.209673194 0.9788095  -1.49100439  2.4713050
#> 38   9.26897317 20.76589728 -0.057478326 0.7747041  -1.59001671  1.4499854
#> 39  -0.39864284 90.66014078  2.953927038 0.7270952   1.54309539  4.3323737
#> 40  28.87602461 61.17786434  5.048191736 0.7970625   3.47987972  6.5924990
#> 41  19.54642753 37.95592405  4.985634749 0.9517308   3.45557409  7.1112137
#> 42  14.88388356 43.57715850 -1.919498702 0.7443177  -3.17265046 -0.2607142
#> 43  12.60139424  3.74310329  3.679018153 0.9767347   1.97894924  5.9412586
#> 44   0.09364355 97.35399138  5.258518384 1.0306855   3.29355651  7.2193326
#> 45  12.14440693 43.17512489 -4.498523515 0.7572986  -5.77629816 -2.8182273
#> 46  -8.59470228 95.75765966 -3.862021119 0.9470826  -5.68283634 -2.0576846
#> 47  -8.28205283 88.77549055 -5.439736492 0.6738603  -6.76359819 -4.1279803
#> 48  21.16208248 63.99787695 -3.791218861 0.9045444  -5.59213899 -2.0494814
#> 49  -5.09940199 97.09666104 -0.004197232 1.0168804  -1.93765281  1.9295129
#> 50  28.55472773 61.88382073  4.445147294 0.8226124   2.81260898  6.0349459
#> 51  16.47522207 33.34272113  3.742561099 1.1215631   1.91361022  6.2041049
#> 52  16.51641676 34.67482482  3.259709769 1.0784461   1.50331896  5.6676081
#> 53  10.91143598 39.84854114 -4.403853199 0.8782799  -5.83426644 -2.4629427
#> 54   5.70007209 78.46927757  5.752066096 0.7964187   4.20625715  7.3310494
#> 55   7.33290636  3.89364911 -1.589517034 0.9765977  -3.28953864  0.6727708
#> 56  -0.55385247 74.87953862 -1.473750746 0.9646297  -3.33134925  0.4240922
#> 57  27.17481798 67.72768302  0.732949152 1.0640290  -1.44361584  2.7799393
#> 58   9.51059931 17.12643304  0.443492017 0.8132455  -1.12948076  2.0076912
#> 59  10.75277214 26.10879638  0.456758575 0.9599662  -1.51253870  2.1543335
#> 60  21.19729056 51.44129347  1.255328057 0.6037163   0.11020277  2.4689874
#> 61  28.67999352 67.56072745  2.304756744 1.0565745   0.15335409  4.3357619
#> 62  -5.22405305 98.28171979  0.191996620 1.0813721  -1.88638658  2.2550802
#> 63  -3.13262629 75.95442676 -3.761507450 0.9105813  -5.52290481 -1.9630739
#> 64  21.30755268 56.64884241 -0.712745248 0.6631426  -1.97428258  0.5451028
#> 65  -9.63819069 84.96897186 -7.826457956 0.6362753  -9.08858631 -6.6068154
#> 66   8.47066509 18.94739354 -0.696451252 0.7811543  -2.23683550  0.7955476
#> 67   7.94939897 27.12866147 -2.622000471 1.0162683  -4.63547193 -0.8355623
#> 68   8.24439430 82.81584852  9.473186611 0.6610972   8.16467397 10.7613465
#> 69  24.21177609 69.32048204 -2.865779631 1.1363272  -5.21483532 -0.6884948
#> 70  10.54902487 24.05447396  0.725183065 0.8597300  -1.07659289  2.3223120
#> 71   4.02549973  4.29887960 -4.897088368 0.9760904  -6.59694528 -2.6346359
#> 72   4.11825703 14.04790941 -4.869318343 0.8677659  -6.50418797 -3.1593126
#> 73  10.49880954 21.63854151  1.070890841 0.7867739  -0.50053832  2.5760872
#> 74  14.98337167 47.93985642 -3.561165539 0.6354550  -4.72249382 -2.2485439
#> 75   9.99270954 19.74103423  0.763753194 0.7737628  -0.77424111  2.2698904
#> 76  -2.10015228 71.93558377 -3.817102097 1.1240425  -6.01708654 -1.5855028
#> 77   7.54531357  0.78847387 -1.376464740 0.9787673  -3.07713143  0.8851780
#> 78   5.67578685 37.54899646 -8.722939395 0.9676129 -10.27814043 -6.5576610
#> 79  15.82139361 51.44077083 -4.120360310 0.6037167  -5.26549438 -2.9066540
#> 80  10.71806576  0.15705542  1.796334561 0.9789504   0.09562076  4.0579302
#> 81  26.35068250 58.16040025  3.727121073 0.7011198   2.38109621  5.0482041
#> 82   8.02849059 15.79052082 -0.993975440 0.8395113  -2.59517251  0.6488476
#> 83  12.95166683 35.90283059 -0.792274071 1.0313888  -2.45745768  1.5153285
#> 84  31.77315249 64.56318784  6.594235293 0.9276361   4.73259646  8.3745523
#> 85   4.24295572 77.58233626  4.054817664 0.8344184   2.43344941  5.6965873
#> 86  18.79388700 56.36468416 -3.113003222 0.6569044  -4.36517789 -1.8482372
#> 87   9.53072176 23.37033986 -0.163373897 0.8337398  -1.89133195  1.4178703
#> 88  14.80599360  8.99805163  5.868957395 0.9407236   4.18354860  7.9271899
#> 89   8.12108168  8.56120649 -0.813354413 0.9464939  -2.50136333  1.2876274
#> 90  10.05104053 30.52183695 -1.615738347 1.1477596  -3.60378500  0.6094368
#> 91  28.02689654 66.74265147  1.978154302 1.0204251  -0.08132140  3.9363905
#> 92   6.44928393  0.02388966 -2.472447275 0.9789504  -4.17316108 -0.2108517
#> 93   8.22126398 20.85699569 -1.114891388 0.7754953  -2.64057931  0.3988228
#> 94  -4.77846206 93.30341273 -0.710248058 0.8302308  -2.31558844  0.8696981
#> 95  -6.16837037 92.56447486 -2.300217710 0.7986215  -3.85315465 -0.7786527
#> 96   3.77165807 73.40943010  2.453740020 1.0424197   0.44600337  4.4816246
#> 97   9.72962780 33.30719834 -2.989161397 1.1225301  -4.82021890 -0.5253337
#> 98  19.02049036 51.50633298 -0.947429451 0.6036780  -2.09146143  0.2648874
#> 99   1.90795738 74.39746463  0.857541544 0.9896860  -1.04981900  2.7889674
#> 100 21.73248835 61.91592400 -2.389904513 0.8237979  -4.02671509 -0.7971306
log_lik(demo_fit)  # Log-likelihood at each demo_fit$data
#>        response        time    loglik      error      Q2.5     Q97.5
#> 1   -3.00841981 91.48060435 -2.252647 0.08344644 -2.427324 -2.102370
#> 2   -7.87686403 93.70754133 -2.758725 0.23743086 -3.288928 -2.382092
#> 3   16.30291008 28.61395348 -3.311628 0.44107430 -4.252498 -2.621291
#> 4   -0.03735530 83.04476261 -2.293959 0.09335656 -2.496819 -2.131667
#> 5   27.44631853 64.17455189 -2.468481 0.17506070 -2.868569 -2.202605
#> 6   22.06100045 51.90959491 -2.372074 0.10682130 -2.603648 -2.192928
#> 7    1.98538897 73.65883146 -2.277995 0.10781077 -2.531637 -2.103686
#> 8   10.35933155 13.46665972 -2.319411 0.11903850 -2.588817 -2.131006
#> 9   15.87725419 65.69922904 -5.829042 0.85849692 -7.622749 -4.308735
#> 10   1.03823613 70.50647840 -2.315816 0.13712836 -2.665991 -2.115741
#> 11  16.41815024 45.77417762 -2.296747 0.08869527 -2.480564 -2.134771
#> 12   0.35867723 71.91122517 -2.336366 0.14403888 -2.693236 -2.130400
#> 13  -2.36615003 93.46722472 -2.359940 0.12639105 -2.639153 -2.157201
#> 14  15.59894731 25.54288243 -3.365739 0.38601527 -4.178997 -2.692371
#> 15  15.20547289 46.22928225 -2.501559 0.13216854 -2.768857 -2.261837
#> 16   0.40988007 94.00145228 -3.064241 0.30521045 -3.720777 -2.561215
#> 17  -4.22113609 97.82264284 -2.303854 0.11868696 -2.584211 -2.118664
#> 18  14.15402439 11.74873617 -3.261354 0.36469190 -4.120859 -2.689911
#> 19  22.43276835 47.49970816 -2.852507 0.19895739 -3.311905 -2.537273
#> 20  25.90014996 56.03327462 -2.871152 0.19980988 -3.289856 -2.534671
#> 21  -8.25310350 90.40313873 -3.164750 0.27470622 -3.755755 -2.703357
#> 22   9.63925445 13.87101677 -2.263641 0.09249695 -2.468900 -2.101645
#> 23  -3.28376193 98.88917289 -2.463780 0.20227007 -2.947434 -2.174843
#> 24  -8.74745808 94.66682326 -2.944851 0.29150990 -3.589196 -2.476863
#> 25   7.82868474  8.24375581 -2.298394 0.10537916 -2.535212 -2.121253
#> 26  23.03457521 51.42117843 -2.592087 0.14395272 -2.917597 -2.353223
#> 27  17.58288831 39.02034671 -2.502406 0.20192543 -3.017927 -2.234914
#> 28  -2.25969226 90.57381309 -2.281771 0.09220271 -2.475748 -2.118513
#> 29  13.80537622 44.69696281 -2.684178 0.17612396 -3.032791 -2.357295
#> 30  -7.11920879 83.60042600 -3.445003 0.29498601 -4.108849 -2.932595
#> 31   5.29891568 73.75956178 -2.878460 0.31101145 -3.568360 -2.386354
#> 32  -1.18941708 81.10551413 -2.244427 0.08062811 -2.413180 -2.094586
#> 33  14.75917506 38.81082828 -2.251265 0.08697731 -2.431718 -2.100241
#> 34  28.77490032 68.51697294 -2.415050 0.18184213 -2.850986 -2.161064
#> 35   5.22268442  0.39483388 -2.767942 0.25333279 -3.301195 -2.315012
#> 36  -0.21033401 83.29160803 -2.284245 0.09072933 -2.477735 -2.125073
#> 37   9.13144062  0.73341469 -2.256321 0.10169475 -2.483722 -2.097216
#> 38   9.26897317 20.76589728 -2.241716 0.08085783 -2.408343 -2.092013
#> 39  -0.39864284 90.66014078 -2.566259 0.16357481 -2.924088 -2.292500
#> 40  28.87602461 61.17786434 -3.195417 0.30277761 -3.843769 -2.674609
#> 41  19.54642753 37.95592405 -3.183754 0.37733961 -4.115606 -2.665950
#> 42  14.88388356 43.57715850 -2.378613 0.11291011 -2.609972 -2.173529
#> 43  12.60139424  3.74310329 -2.761911 0.29490470 -3.492994 -2.367343
#> 44   0.09364355 97.35399138 -3.296765 0.41466638 -4.217226 -2.626155
#> 45  12.14440693 43.17512489 -3.000742 0.24906993 -3.487285 -2.529142
#> 46  -8.59470228 95.75765966 -2.811124 0.27727574 -3.423020 -2.377289
#> 47  -8.28205283 88.77549055 -3.346031 0.28846486 -3.976048 -2.861096
#> 48  21.16208248 63.99787695 -2.790860 0.25884624 -3.368785 -2.385983
#> 49  -5.09940199 97.09666104 -2.257613 0.08959579 -2.447822 -2.102607
#> 50  28.55472773 61.88382073 -2.982569 0.27262733 -3.566072 -2.522542
#> 51  16.47522207 33.34272113 -2.790974 0.33949459 -3.662880 -2.355943
#> 52  16.51641676 34.67482482 -2.660404 0.28950495 -3.418479 -2.290439
#> 53  10.91143598 39.84854114 -2.976107 0.27680447 -3.510285 -2.453217
#> 54   5.70007209 78.46927757 -3.481132 0.35113818 -4.240148 -2.871162
#> 55   7.33290636  3.89364911 -2.349348 0.12526154 -2.633156 -2.146193
#> 56  -0.55385247 74.87953862 -2.335629 0.12983097 -2.647213 -2.139672
#> 57  27.17481798 67.72768302 -2.280600 0.10910635 -2.547264 -2.105851
#> 58   9.51059931 17.12643304 -2.251594 0.08304033 -2.421010 -2.093650
#> 59  10.75277214 26.10879638 -2.261598 0.08600253 -2.440759 -2.107126
#> 60  21.19729056 51.44129347 -2.291482 0.09133065 -2.486391 -2.133029
#> 61  28.67999352 67.56072745 -2.457971 0.19350730 -2.911148 -2.180797
#> 62  -5.22405305 98.28171979 -2.264019 0.09397048 -2.468894 -2.101872
#> 63  -3.13262629 75.95442676 -2.782034 0.25669158 -3.366335 -2.376162
#> 64  21.30755268 56.64884241 -2.255100 0.08071503 -2.422441 -2.101633
#> 65  -9.63819069 84.96897186 -4.532481 0.45836924 -5.486792 -3.731865
#> 66   8.47066509 18.94739354 -2.259836 0.08905653 -2.458879 -2.106296
#> 67   7.94939897 27.12866147 -2.514180 0.21491792 -3.021808 -2.213943
#> 68   8.24439430 82.81584852 -5.598296 0.61633576 -6.912699 -4.462386
#> 69  24.21177609 69.32048204 -2.576789 0.25492110 -3.218774 -2.217935
#> 70  10.54902487 24.05447396 -2.266944 0.08805570 -2.452656 -2.105119
#> 71   4.02549973  4.29887960 -3.153940 0.34344720 -3.872773 -2.495062
#> 72   4.11825703 14.04790941 -3.135276 0.31758148 -3.824826 -2.603905
#> 73  10.49880954 21.63854151 -2.286080 0.09508665 -2.495017 -2.122743
#> 74  14.98337167 47.93985642 -2.711249 0.16577478 -3.045324 -2.412843
#> 75   9.99270954 19.74103423 -2.264016 0.08693118 -2.451161 -2.108140
#> 76  -2.10015228 71.93558377 -2.814185 0.32326936 -3.556260 -2.328175
#> 77   7.54531357  0.78847387 -2.325792 0.11616692 -2.588514 -2.133859
#> 78   5.67578685 37.54899646 -5.108715 0.70631183 -6.495399 -3.776135
#> 79  15.82139361 51.44077083 -2.871317 0.18689401 -3.259842 -2.537136
#> 80  10.71806576  0.15705542 -2.375596 0.17359328 -2.842370 -2.153396
#> 81  26.35068250 58.16040025 -2.756438 0.19513914 -3.171350 -2.430585
#> 82   8.02849059 15.79052082 -2.282281 0.09868108 -2.497267 -2.113488
#> 83  12.95166683 35.90283059 -2.282271 0.09577160 -2.495318 -2.114439
#> 84  31.77315249 64.56318784 -3.877298 0.47393906 -4.895958 -3.042466
#> 85   4.24295572 77.58233626 -2.859787 0.25107469 -3.414302 -2.444844
#> 86  18.79388700 56.36468416 -2.600350 0.15336269 -2.927620 -2.340919
#> 87   9.53072176 23.37033986 -2.245975 0.08423112 -2.419778 -2.093793
#> 88  14.80599360  8.99805163 -3.543508 0.44116725 -4.589128 -2.867354
#> 89   8.12108168  8.56120649 -2.277212 0.09664895 -2.497941 -2.111224
#> 90  10.05104053 30.52183695 -2.365845 0.15209532 -2.711968 -2.134841
#> 91  28.02689654 66.74265147 -2.403031 0.16730320 -2.799757 -2.160779
#> 92   6.44928393  0.02388966 -2.484033 0.17257075 -2.873297 -2.202031
#> 93   8.22126398 20.85699569 -2.287494 0.10088282 -2.514554 -2.117132
#> 94  -4.77846206 93.30341273 -2.263775 0.09118010 -2.459846 -2.102406
#> 95  -6.16837037 92.56447486 -2.441244 0.14864452 -2.769742 -2.198362
#> 96   3.77165807 73.40943010 -2.483579 0.20218001 -2.966147 -2.189095
#> 97   9.72962780 33.30719834 -2.601103 0.23721271 -3.087431 -2.207647
#> 98  19.02049036 51.50633298 -2.267005 0.08062755 -2.432796 -2.115335
#> 99   1.90795738 74.39746463 -2.282380 0.10892537 -2.538169 -2.107621
#> 100 21.73248835 61.91592400 -2.460010 0.15412517 -2.819519 -2.215585

# All of the above take a range of arguments. E.g.,:
# \donttest{
predict(demo_fit)  # Pointwise posterior predictive
#>            time     predict    error       Q2.5     Q97.5
#> 1   91.48060435 -3.64868144 3.726942 -10.707182  3.717179
#> 2   93.70754133 -4.12695871 3.831065 -11.516781  3.620307
#> 3   28.61395348 10.95164701 3.832366   3.447263 18.401225
#> 4   83.04476261 -1.22990275 3.728680  -8.296429  6.253298
#> 5   64.17455189 25.05926220 3.898731  17.449743 32.676730
#> 6   51.90959491 20.18136827 3.681843  12.929877 27.348292
#> 7   73.65883146  1.19329938 3.802991  -6.337524  8.512163
#> 8   13.46665972  8.90621203 3.757642   1.483056 16.412671
#> 9   65.69922904 25.65457241 3.816836  18.013090 33.000796
#> 10  70.50647840  2.03078966 3.798703  -5.440229  9.386004
#> 11  45.77417762 17.71211382 3.748208  10.410457 25.147970
#> 12  71.91122517  1.72022753 3.864502  -5.527213  9.417072
#> 13  93.46722472 -4.08989565 3.750527 -11.480943  3.144094
#> 14  25.54288243 10.09736558 3.884298   2.613024 17.924775
#> 15  46.22928225 17.92538443 3.629945  10.435705 24.794048
#> 16  94.00145228 -4.17550632 3.766060 -11.693122  3.224694
#> 17  97.82264284 -5.35176787 3.752839 -12.739065  1.896711
#> 18  11.74873617  8.93829798 3.719764   1.627326 16.192727
#> 19  47.49970816 18.40046954 3.735844  11.152365 25.639188
#> 20  56.03327462 21.85467879 3.899195  14.074651 29.404462
#> 21  90.40313873 -3.19444641 3.724294 -10.429835  4.026217
#> 22  13.87101677  8.92539752 3.792989   1.346452 16.359147
#> 23  98.88917289 -5.63609229 3.894419 -13.346249  2.026696
#> 24  94.66682326 -4.36667122 3.759597 -12.002887  2.936692
#> 25   8.24375581  8.89997181 3.865105   1.380347 16.646245
#> 26  51.42117843 19.93774549 3.722389  12.753670 27.363595
#> 27  39.02034671 14.94402978 3.791411   7.523750 22.180542
#> 28  90.57381309 -3.32372637 3.735421 -10.751252  4.028962
#> 29  44.69696281 17.32977392 3.725537   9.905396 24.706880
#> 30  83.60042600 -1.47739339 3.729916  -8.602392  5.979920
#> 31  73.75956178  1.25197174 3.899441  -6.681165  8.875523
#> 32  81.10551413 -0.83837553 3.757759  -8.366212  6.596885
#> 33  38.81082828 14.81204319 3.840117   7.180708 22.545080
#> 34  68.51697294 26.80244828 3.829993  19.529936 34.217401
#> 35   0.39483388  8.92473499 3.852978   1.165191 16.399109
#> 36  83.29160803 -1.34087762 3.786490  -8.729799  5.949986
#> 37   0.73341469  8.85992336 3.887000   1.307223 16.582029
#> 38  20.76589728  9.25529068 3.771079   1.658375 16.608109
#> 39  90.66014078 -3.41958467 3.812491 -10.992376  4.019831
#> 40  61.17786434 23.79922202 3.832847  16.266375 31.176195
#> 41  37.95592405 14.63035528 3.812356   7.304104 22.001430
#> 42  43.57715850 16.77164220 3.767679   9.402033 23.930636
#> 43   3.74310329  8.87985018 3.872041   1.062761 16.235017
#> 44  97.35399138 -5.31261368 3.826167 -13.003495  1.911415
#> 45  43.17512489 16.61997678 3.785464   9.185825 23.973077
#> 46  95.75765966 -4.69784217 3.887609 -12.261589  2.805238
#> 47  88.77549055 -2.86396088 3.775418 -10.156431  4.636638
#> 48  63.99787695 24.85808960 3.803667  17.358088 32.334752
#> 49  97.09666104 -5.07924317 3.807903 -12.567566  2.394021
#> 50  61.88382073 23.96930858 3.848568  16.365472 31.392866
#> 51  33.34272113 12.62791560 3.995353   4.954471 20.569222
#> 52  34.67482482 13.22928971 3.854728   5.765818 21.030835
#> 53  39.84854114 15.40009244 3.789223   8.041966 22.972831
#> 54  78.46927757 -0.01810456 3.811762  -7.417191  7.258544
#> 55   3.89364911  8.95925582 3.823272   1.087624 16.338376
#> 56  74.87953862  0.97353814 3.829937  -6.597649  8.692553
#> 57  67.72768302 26.47613893 3.879408  18.811379 33.972008
#> 58  17.12643304  9.06024726 3.772427   1.656259 16.448662
#> 59  26.10879638 10.30948331 3.829177   2.764366 18.045593
#> 60  51.44129347 19.99063286 3.763778  12.576646 27.332112
#> 61  67.56072745 26.30091463 3.888796  18.551904 34.018078
#> 62  98.28171979 -5.38967238 3.815650 -13.012466  2.124523
#> 63  75.95442676  0.65867540 3.742343  -6.468769  8.081778
#> 64  56.64884241 22.09982088 3.682438  14.848067 29.492631
#> 65  84.96897186 -1.87722756 3.753484  -9.252855  5.459104
#> 66  18.94739354  9.12444784 3.860966   1.598001 16.873361
#> 67  27.12866147 10.53953037 3.899364   3.015452 18.283395
#> 68  82.81584852 -1.05848464 3.755892  -8.470051  6.097112
#> 69  69.32048204 27.15447541 3.850827  19.326703 34.638033
#> 70  24.05447396  9.98139676 3.740818   2.764461 17.249538
#> 71   4.29887960  8.86401231 3.853403   1.335521 16.515382
#> 72  14.04790941  8.95198369 3.774944   1.413434 16.228605
#> 73  21.63854151  9.45007369 3.702680   2.181947 16.564276
#> 74  47.93985642 18.58204223 3.703971  11.290536 26.186875
#> 75  19.74103423  9.22038096 3.801558   1.873347 16.644345
#> 76  71.93558377  1.67506270 3.871108  -5.944599  9.566824
#> 77   0.78847387  8.88539128 3.772671   1.471900 16.099317
#> 78  37.54899646 14.36943569 3.878165   6.859171 21.991207
#> 79  51.44077083 19.86591580 3.611697  12.394238 27.082061
#> 80   0.15705542  8.84273146 3.811527   1.436537 16.433420
#> 81  58.16040025 22.68805422 3.783699  15.149422 29.917763
#> 82  15.79052082  9.01782710 3.748765   1.597716 16.412802
#> 83  35.90283059 13.71852955 3.833460   6.114017 21.402819
#> 84  64.56318784 25.14816081 3.724130  17.795019 32.563745
#> 85  77.58233626  0.26227737 3.679990  -6.994186  7.411076
#> 86  56.36468416 21.95172535 3.688631  14.925786 29.060793
#> 87  23.37033986  9.72997990 3.796028   2.123583 17.255789
#> 88   8.99805163  8.90807543 3.837715   1.351269 16.643081
#> 89   8.56120649  8.95315259 3.749876   1.341656 16.357022
#> 90  30.52183695 11.74084856 3.926752   4.052797 19.575215
#> 91  66.74265147 25.98977266 3.836373  18.414582 33.395888
#> 92   0.02388966  8.84505334 3.858287   1.207938 16.632909
#> 93  20.85699569  9.28430464 3.838182   1.594536 17.015631
#> 94  93.30341273 -4.13063479 3.807830 -11.743512  3.301895
#> 95  92.56447486 -3.83826152 3.840764 -11.360827  3.876434
#> 96  73.40943010  1.35028668 3.868471  -6.271649  8.736902
#> 97  33.30719834 12.73740627 3.908295   4.999705 20.280296
#> 98  51.50633298 19.93902061 3.758074  12.647858 27.363301
#> 99  74.39746463  1.04077985 3.843613  -6.308708  8.660525
#> 100 61.91592400 24.20386944 3.752874  16.746924 31.645314
predict(demo_fit, probs = c(0.1, 0.5, 0.9))  # With median and 80% credible interval.
#>            time    predict    error        Q10         Q50        Q90
#> 1   91.48060435 -3.5424945 3.774480  -8.428804 -3.50713312  1.2400256
#> 2   93.70754133 -4.2340457 3.784336  -8.945190 -4.22840407  0.5838814
#> 3   28.61395348 10.8697200 3.830395   6.038887 10.77935811 15.7917445
#> 4   83.04476261 -1.2926620 3.825903  -6.111511 -1.21860326  3.5483828
#> 5   64.17455189 25.1876293 3.789109  20.390234 25.14558737 30.0624137
#> 6   51.90959491 20.0857404 3.737858  15.399125 20.04328520 24.9159403
#> 7   73.65883146  1.3299629 3.893323  -3.686766  1.34998334  6.2670500
#> 8   13.46665972  8.8915413 3.715494   4.041510  8.97143670 13.5880057
#> 9   65.69922904 25.5395004 3.847843  20.634190 25.54671861 30.4007333
#> 10  70.50647840  1.9922636 3.856852  -2.946644  2.08666663  6.8834588
#> 11  45.77417762 17.6748472 3.712330  12.972358 17.66470518 22.4724899
#> 12  71.91122517  1.6652049 3.848410  -3.254839  1.65466461  6.5882638
#> 13  93.46722472 -4.0086869 3.770194  -8.696752 -4.06095359  0.7734857
#> 14  25.54288243 10.1152250 3.822152   5.229437 10.01225508 15.1482170
#> 15  46.22928225 17.9784577 3.756859  13.211428 18.07110666 22.7989873
#> 16  94.00145228 -4.1819124 3.848294  -9.286448 -4.29163788  0.7186379
#> 17  97.82264284 -5.3545472 3.814376 -10.039766 -5.37826694 -0.4886659
#> 18  11.74873617  9.0192907 3.840544   4.219160  9.00815121 13.8973493
#> 19  47.49970816 18.3759160 3.664630  13.792566 18.34331555 23.0560704
#> 20  56.03327462 21.6840238 3.827511  16.639201 21.67727573 26.6326441
#> 21  90.40313873 -3.2749157 3.785848  -8.113411 -3.31786680  1.6274316
#> 22  13.87101677  8.9407037 3.758123   4.190074  8.96801173 13.7307075
#> 23  98.88917289 -5.6025631 3.907870 -10.638293 -5.58501717 -0.5552101
#> 24  94.66682326 -4.4533497 3.781063  -9.185813 -4.45648016  0.3611750
#> 25   8.24375581  9.0474232 3.861365   4.187596  8.96885604 14.0209838
#> 26  51.42117843 19.8484994 3.810406  14.856481 19.92025793 24.6860260
#> 27  39.02034671 14.9946638 3.826325  10.116795 15.01585646 19.9250662
#> 28  90.57381309 -3.1843376 3.744292  -7.988644 -3.17271758  1.6071091
#> 29  44.69696281 17.3131006 3.772043  12.607148 17.26961867 22.0578492
#> 30  83.60042600 -1.4809200 3.788183  -6.374664 -1.41276614  3.4582215
#> 31  73.75956178  1.1981907 3.793292  -3.492611  1.20435446  6.1159661
#> 32  81.10551413 -0.7076087 3.735970  -5.386280 -0.75046634  4.1799531
#> 33  38.81082828 14.9367752 3.766341  10.112990 14.85998462 19.9026235
#> 34  68.51697294 26.6249541 3.869297  21.771607 26.63001992 31.6375390
#> 35   0.39483388  8.8552919 3.828190   3.970532  8.83936891 13.7244791
#> 36  83.29160803 -1.3540967 3.714495  -6.053644 -1.34152106  3.2708044
#> 37   0.73341469  8.8481473 3.782391   4.071572  8.93362534 13.6797615
#> 38  20.76589728  9.2338334 3.798666   4.298846  9.19348632 14.1260340
#> 39  90.66014078 -3.4676744 3.712572  -8.193945 -3.47748768  1.1626070
#> 40  61.17786434 23.7426045 3.880912  18.890116 23.73677279 28.7722394
#> 41  37.95592405 14.5658225 3.826241   9.811088 14.52709057 19.3905651
#> 42  43.57715850 16.8526812 3.748629  11.949907 16.81949131 21.6546963
#> 43   3.74310329  8.8826049 3.869974   3.988288  8.89414367 13.8051554
#> 44  97.35399138 -5.2145256 3.817948 -10.122435 -5.18524939 -0.3689203
#> 45  43.17512489 16.7403726 3.770011  11.846707 16.77979954 21.6388649
#> 46  95.75765966 -4.7514596 3.849642  -9.698557 -4.74912026  0.2966069
#> 47  88.77549055 -2.8145685 3.726580  -7.663892 -2.76462856  1.9225459
#> 48  63.99787695 24.9719651 3.862312  19.964725 24.99891736 29.8843444
#> 49  97.09666104 -5.0949747 3.807036 -10.004457 -5.16720553 -0.1980743
#> 50  61.88382073 24.1583913 3.726670  19.383535 24.10979210 28.9415826
#> 51  33.34272113 12.8328980 3.781243   7.997075 12.78331497 17.7968502
#> 52  34.67482482 13.2780485 3.884796   8.268969 13.31565199 18.2382064
#> 53  39.84854114 15.2192350 3.769034  10.387806 15.22197677 20.0736601
#> 54  78.46927757 -0.1558587 3.730506  -5.001270 -0.11981125  4.5776622
#> 55   3.89364911  8.9878521 3.806108   4.114069  8.96142690 13.8731434
#> 56  74.87953862  0.7864903 3.805431  -4.104981  0.79676871  5.6706912
#> 57  67.72768302 26.4206799 3.762597  21.713774 26.37119045 31.1057527
#> 58  17.12643304  9.1686805 3.750581   4.311651  9.31709976 13.8925735
#> 59  26.10879638 10.3347196 3.844350   5.297995 10.33393847 15.2361269
#> 60  51.44129347 19.9551304 3.811192  15.161019 19.85284447 24.8788750
#> 61  67.56072745 26.2977902 3.844677  21.487063 26.32271978 31.2187580
#> 62  98.28171979 -5.4092119 3.843567 -10.257119 -5.40717975 -0.5068134
#> 63  75.95442676  0.6919126 3.774904  -3.986208  0.63030068  5.5786457
#> 64  56.64884241 22.1635863 3.775014  17.281214 22.13675479 27.2395316
#> 65  84.96897186 -1.9075976 3.754488  -6.659835 -1.84900705  2.8016680
#> 66  18.94739354  9.1183455 3.742413   4.308182  9.11977331 13.9543859
#> 67  27.12866147 10.6039352 3.855878   5.711578 10.62268337 15.4833715
#> 68  82.81584852 -1.1891589 3.700322  -5.675901 -1.27541646  3.6348593
#> 69  69.32048204 27.1422356 3.792583  22.343893 27.15480518 31.8838834
#> 70  24.05447396  9.8364514 3.879891   4.895085  9.78972961 14.9087187
#> 71   4.29887960  8.8716886 3.925185   3.834384  8.86606049 13.7963896
#> 72  14.04790941  9.0150692 3.826937   4.327242  8.94267623 13.8014149
#> 73  21.63854151  9.5094308 3.782453   4.789779  9.49801197 14.2262623
#> 74  47.93985642 18.5150344 3.693633  13.821613 18.58936279 23.2555400
#> 75  19.74103423  9.2583755 3.749897   4.498498  9.26229972 13.9958640
#> 76  71.93558377  1.8136066 3.838969  -3.098719  1.78448787  6.7609602
#> 77   0.78847387  8.9261178 3.732295   4.112124  8.87582002 13.7195524
#> 78  37.54899646 14.2289392 3.833558   9.242378 14.25463820 19.1683747
#> 79  51.44077083 19.9133763 3.784604  15.105985 19.89981395 24.7138017
#> 80   0.15705542  8.9138764 3.884051   3.901332  8.97712424 13.7627841
#> 81  58.16040025 22.7381509 3.790769  18.082650 22.64132584 27.5635417
#> 82  15.79052082  8.9128659 3.801156   3.915292  8.95373881 13.7944812
#> 83  35.90283059 13.8064412 3.852080   9.014005 13.79981053 18.6564387
#> 84  64.56318784 25.1493953 3.883144  20.175697 25.08006296 30.0444024
#> 85  77.58233626  0.1455059 3.730553  -4.637601  0.07272973  4.9509381
#> 86  56.36468416 21.9069505 3.735490  17.252128 21.89464193 26.7069082
#> 87  23.37033986  9.6851937 3.743637   4.805468  9.77921930 14.4024909
#> 88   8.99805163  8.9999542 3.852335   4.092822  8.94526199 13.9261721
#> 89   8.56120649  8.9975623 3.832749   4.119161  8.99245390 13.8639141
#> 90  30.52183695 11.7025478 3.871570   6.784152 11.79632655 16.6773016
#> 91  66.74265147 26.0523513 3.823974  21.033791 26.09382613 30.9712698
#> 92   0.02388966  8.9283668 3.806543   4.099252  8.88293985 13.8470813
#> 93  20.85699569  9.3704237 3.773850   4.613868  9.35500115 14.2721421
#> 94  93.30341273 -3.9885989 3.663280  -8.621433 -4.01146580  0.6519968
#> 95  92.56447486 -3.7946715 3.795457  -8.641084 -3.80838371  1.0938448
#> 96  73.40943010  1.4015989 3.792254  -3.399549  1.39361322  6.2101812
#> 97  33.30719834 12.6877410 3.902726   7.683689 12.66500161 17.5671020
#> 98  51.50633298 20.0149853 3.809823  15.139184 20.09830966 24.8856429
#> 99  74.39746463  1.0288193 3.900601  -3.868948  1.03060169  6.0709719
#> 100 61.91592400 24.1068237 3.779167  19.350453 24.15330672 28.9487542
predict(demo_fit, prior = TRUE)  # Prior predictive
#>            time   predict    error      Q2.5    Q97.5
#> 1   91.48060435  9.635897 36.30737 -62.15975 80.20683
#> 2   93.70754133  9.264177 36.89372 -61.10685 80.47293
#> 3   28.61395348 10.039237 29.22904 -41.77453 61.76280
#> 4   83.04476261  9.103839 38.64471 -55.26422 77.96218
#> 5   64.17455189  8.414077 33.95355 -54.92302 67.73977
#> 6   51.90959491  9.578506 32.05452 -46.19895 70.05253
#> 7   73.65883146  9.481321 36.76180 -54.65116 79.73796
#> 8   13.46665972  8.920732 25.80729 -43.47148 59.31483
#> 9   65.69922904  8.534149 34.62219 -53.05431 71.27394
#> 10  70.50647840  8.362193 36.02394 -53.22391 71.63913
#> 11  45.77417762  9.195096 31.00748 -47.94419 64.66657
#> 12  71.91122517  9.055669 37.16067 -54.31152 72.88903
#> 13  93.46722472  9.873370 36.18826 -61.01616 85.49374
#> 14  25.54288243  9.511140 26.45631 -45.38311 65.10878
#> 15  46.22928225  9.481099 30.95649 -44.88065 68.23072
#> 16  94.00145228 10.097045 36.21021 -59.80715 81.08277
#> 17  97.82264284 10.161640 38.38215 -65.52482 83.27567
#> 18  11.74873617  9.143410 26.31430 -45.13748 60.33671
#> 19  47.49970816  9.502943 30.63812 -47.16249 66.13707
#> 20  56.03327462  9.256709 32.23412 -48.77521 70.14558
#> 21  90.40313873  9.851856 36.51906 -61.37215 84.33408
#> 22  13.87101677  9.265833 26.04009 -43.89526 60.13762
#> 23  98.88917289 10.115373 38.90251 -63.34880 89.13085
#> 24  94.66682326  9.889975 37.59861 -63.37048 84.32840
#> 25   8.24375581  9.612430 26.26758 -38.99532 62.92046
#> 26  51.42117843  9.155364 31.49711 -45.57331 66.94382
#> 27  39.02034671  9.276091 30.00310 -47.50144 62.17046
#> 28  90.57381309  9.707100 36.51456 -67.37705 81.97005
#> 29  44.69696281  9.267766 30.37585 -47.97701 64.94715
#> 30  83.60042600  9.451294 38.99110 -58.06650 79.42831
#> 31  73.75956178  8.579979 36.76168 -56.64539 71.42570
#> 32  81.10551413  9.202515 38.31633 -53.61800 74.73897
#> 33  38.81082828  9.671662 29.55472 -46.90162 65.31809
#> 34  68.51697294  8.911829 35.65210 -55.47824 71.67754
#> 35   0.39483388  9.284125 27.86157 -45.51696 58.93856
#> 36  83.29160803  9.979204 39.31900 -54.70081 80.13777
#> 37   0.73341469  9.439961 26.03139 -40.80184 60.50905
#> 38  20.76589728  8.975016 26.56389 -46.25777 59.85300
#> 39  90.66014078  9.606229 36.30575 -62.48921 82.37112
#> 40  61.17786434  9.219674 33.25039 -51.55550 69.25959
#> 41  37.95592405  9.281724 28.47074 -44.44359 62.10226
#> 42  43.57715850  8.887878 31.51894 -48.48793 64.22978
#> 43   3.74310329  9.129714 26.71436 -40.34802 63.08278
#> 44  97.35399138 10.360299 37.61599 -61.54227 86.28546
#> 45  43.17512489  8.978397 30.56774 -47.36683 65.05760
#> 46  95.75765966 10.304975 37.42218 -62.71051 87.72117
#> 47  88.77549055  9.081950 36.22157 -61.17460 80.07615
#> 48  63.99787695  8.357851 34.18711 -56.37686 69.29575
#> 49  97.09666104 10.675580 37.54181 -63.08688 83.11350
#> 50  61.88382073  8.272680 33.12024 -51.76040 67.26074
#> 51  33.34272113  8.925765 28.63162 -44.17310 62.39947
#> 52  34.67482482  9.172597 28.79940 -44.65654 62.60379
#> 53  39.84854114  8.819916 28.83203 -44.24736 62.17213
#> 54  78.46927757  8.645662 37.66164 -56.77097 74.07225
#> 55   3.89364911  9.476324 26.13114 -43.14137 60.21846
#> 56  74.87953862  9.350426 36.70407 -52.14865 72.53215
#> 57  67.72768302  8.809771 34.15633 -54.60982 72.06005
#> 58  17.12643304 10.018477 26.78179 -41.29686 61.73471
#> 59  26.10879638  9.660509 27.44009 -45.08404 64.41534
#> 60  51.44129347  9.500152 31.28776 -48.52949 68.30627
#> 61  67.56072745  9.390220 34.30565 -51.67714 71.35073
#> 62  98.28171979  9.896879 38.05815 -65.50340 89.71044
#> 63  75.95442676  8.945299 36.96879 -56.45700 74.84699
#> 64  56.64884241  9.275983 32.33656 -50.04984 69.81424
#> 65  84.96897186  9.504117 39.81358 -55.50099 78.78632
#> 66  18.94739354  9.259361 26.29153 -43.52013 61.02938
#> 67  27.12866147  9.733971 28.66051 -43.59321 63.78401
#> 68  82.81584852  9.140648 39.74973 -57.70018 81.72176
#> 69  69.32048204  8.750302 35.27190 -52.50134 72.62262
#> 70  24.05447396  9.849707 26.49256 -41.31999 63.46407
#> 71   4.29887960  9.027624 26.23468 -45.09882 62.04223
#> 72  14.04790941  9.295145 25.33399 -42.88505 61.57859
#> 73  21.63854151  9.421524 25.68756 -41.62616 62.25753
#> 74  47.93985642  8.994598 31.20273 -51.04679 65.45653
#> 75  19.74103423  8.551137 25.52567 -47.62570 58.14894
#> 76  71.93558377  9.113677 35.28501 -50.36161 72.56972
#> 77   0.78847387  8.597248 27.65651 -47.06289 60.46776
#> 78  37.54899646  9.576958 28.87606 -45.09062 65.04069
#> 79  51.44077083  9.527177 31.14950 -45.99304 67.55829
#> 80   0.15705542  8.976311 25.15951 -41.87904 59.04946
#> 81  58.16040025  8.774905 32.42278 -49.16729 66.99882
#> 82  15.79052082  9.047140 26.26670 -43.28971 58.95499
#> 83  35.90283059  9.311575 28.41940 -43.34206 61.70849
#> 84  64.56318784  8.376573 33.91338 -52.90827 68.33797
#> 85  77.58233626  9.502982 36.96815 -55.10132 75.44002
#> 86  56.36468416  9.083765 32.37078 -48.38130 68.24804
#> 87  23.37033986  9.642091 27.00756 -42.21017 63.79313
#> 88   8.99805163  9.071151 26.04038 -43.98776 60.19581
#> 89   8.56120649  9.402536 26.54183 -44.38189 63.42992
#> 90  30.52183695  9.794209 28.48639 -44.22497 65.35245
#> 91  66.74265147  9.214087 34.10111 -53.11340 69.44213
#> 92   0.02388966  9.351853 25.70629 -39.92834 62.18011
#> 93  20.85699569  8.788919 26.46829 -44.51949 60.01503
#> 94  93.30341273  9.619201 37.17722 -63.32068 83.66034
#> 95  92.56447486  9.442558 36.69394 -63.32525 82.22406
#> 96  73.40943010  9.239871 36.88498 -53.98744 76.00703
#> 97  33.30719834  9.092535 27.47881 -47.50308 60.80199
#> 98  51.50633298  8.796488 31.18649 -49.66702 62.93892
#> 99  74.39746463  9.080934 36.68041 -56.13218 77.21370
#> 100 61.91592400  8.877250 34.00073 -55.44571 71.43347
fitted(demo_fit, summary = FALSE)  # Samples instead of summary. Useful for plotting distributions.
#> # A tibble: 300,000 × 13
#>    .chain .iteration .draw  cp_1  cp_2 Intercept_1 time_2 Intercept_3 time_3
#>     <int>      <int> <int> <dbl> <dbl>       <dbl>  <dbl>       <dbl>  <dbl>
#>  1      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  2      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  3      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  4      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  5      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  6      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  7      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  8      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#>  9      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#> 10      1          1     1  24.4  69.8        8.72  0.412        1.90 -0.231
#> # ℹ 299,990 more rows
#> # ℹ 4 more variables: sigma_1 <dbl>, time <dbl>, data_row <int>, .epred <dbl>
fitted(demo_fit, dpar = "sigma")  # Another model parameter
#>            time   fitted     error     Q2.5    Q97.5
#> 1   91.48060435 3.681777 0.2758695 3.196039 4.297563
#> 2   93.70754133 3.681777 0.2758695 3.196039 4.297563
#> 3   28.61395348 3.681777 0.2758695 3.196039 4.297563
#> 4   83.04476261 3.681777 0.2758695 3.196039 4.297563
#> 5   64.17455189 3.681777 0.2758695 3.196039 4.297563
#> 6   51.90959491 3.681777 0.2758695 3.196039 4.297563
#> 7   73.65883146 3.681777 0.2758695 3.196039 4.297563
#> 8   13.46665972 3.681777 0.2758695 3.196039 4.297563
#> 9   65.69922904 3.681777 0.2758695 3.196039 4.297563
#> 10  70.50647840 3.681777 0.2758695 3.196039 4.297563
#> 11  45.77417762 3.681777 0.2758695 3.196039 4.297563
#> 12  71.91122517 3.681777 0.2758695 3.196039 4.297563
#> 13  93.46722472 3.681777 0.2758695 3.196039 4.297563
#> 14  25.54288243 3.681777 0.2758695 3.196039 4.297563
#> 15  46.22928225 3.681777 0.2758695 3.196039 4.297563
#> 16  94.00145228 3.681777 0.2758695 3.196039 4.297563
#> 17  97.82264284 3.681777 0.2758695 3.196039 4.297563
#> 18  11.74873617 3.681777 0.2758695 3.196039 4.297563
#> 19  47.49970816 3.681777 0.2758695 3.196039 4.297563
#> 20  56.03327462 3.681777 0.2758695 3.196039 4.297563
#> 21  90.40313873 3.681777 0.2758695 3.196039 4.297563
#> 22  13.87101677 3.681777 0.2758695 3.196039 4.297563
#> 23  98.88917289 3.681777 0.2758695 3.196039 4.297563
#> 24  94.66682326 3.681777 0.2758695 3.196039 4.297563
#> 25   8.24375581 3.681777 0.2758695 3.196039 4.297563
#> 26  51.42117843 3.681777 0.2758695 3.196039 4.297563
#> 27  39.02034671 3.681777 0.2758695 3.196039 4.297563
#> 28  90.57381309 3.681777 0.2758695 3.196039 4.297563
#> 29  44.69696281 3.681777 0.2758695 3.196039 4.297563
#> 30  83.60042600 3.681777 0.2758695 3.196039 4.297563
#> 31  73.75956178 3.681777 0.2758695 3.196039 4.297563
#> 32  81.10551413 3.681777 0.2758695 3.196039 4.297563
#> 33  38.81082828 3.681777 0.2758695 3.196039 4.297563
#> 34  68.51697294 3.681777 0.2758695 3.196039 4.297563
#> 35   0.39483388 3.681777 0.2758695 3.196039 4.297563
#> 36  83.29160803 3.681777 0.2758695 3.196039 4.297563
#> 37   0.73341469 3.681777 0.2758695 3.196039 4.297563
#> 38  20.76589728 3.681777 0.2758695 3.196039 4.297563
#> 39  90.66014078 3.681777 0.2758695 3.196039 4.297563
#> 40  61.17786434 3.681777 0.2758695 3.196039 4.297563
#> 41  37.95592405 3.681777 0.2758695 3.196039 4.297563
#> 42  43.57715850 3.681777 0.2758695 3.196039 4.297563
#> 43   3.74310329 3.681777 0.2758695 3.196039 4.297563
#> 44  97.35399138 3.681777 0.2758695 3.196039 4.297563
#> 45  43.17512489 3.681777 0.2758695 3.196039 4.297563
#> 46  95.75765966 3.681777 0.2758695 3.196039 4.297563
#> 47  88.77549055 3.681777 0.2758695 3.196039 4.297563
#> 48  63.99787695 3.681777 0.2758695 3.196039 4.297563
#> 49  97.09666104 3.681777 0.2758695 3.196039 4.297563
#> 50  61.88382073 3.681777 0.2758695 3.196039 4.297563
#> 51  33.34272113 3.681777 0.2758695 3.196039 4.297563
#> 52  34.67482482 3.681777 0.2758695 3.196039 4.297563
#> 53  39.84854114 3.681777 0.2758695 3.196039 4.297563
#> 54  78.46927757 3.681777 0.2758695 3.196039 4.297563
#> 55   3.89364911 3.681777 0.2758695 3.196039 4.297563
#> 56  74.87953862 3.681777 0.2758695 3.196039 4.297563
#> 57  67.72768302 3.681777 0.2758695 3.196039 4.297563
#> 58  17.12643304 3.681777 0.2758695 3.196039 4.297563
#> 59  26.10879638 3.681777 0.2758695 3.196039 4.297563
#> 60  51.44129347 3.681777 0.2758695 3.196039 4.297563
#> 61  67.56072745 3.681777 0.2758695 3.196039 4.297563
#> 62  98.28171979 3.681777 0.2758695 3.196039 4.297563
#> 63  75.95442676 3.681777 0.2758695 3.196039 4.297563
#> 64  56.64884241 3.681777 0.2758695 3.196039 4.297563
#> 65  84.96897186 3.681777 0.2758695 3.196039 4.297563
#> 66  18.94739354 3.681777 0.2758695 3.196039 4.297563
#> 67  27.12866147 3.681777 0.2758695 3.196039 4.297563
#> 68  82.81584852 3.681777 0.2758695 3.196039 4.297563
#> 69  69.32048204 3.681777 0.2758695 3.196039 4.297563
#> 70  24.05447396 3.681777 0.2758695 3.196039 4.297563
#> 71   4.29887960 3.681777 0.2758695 3.196039 4.297563
#> 72  14.04790941 3.681777 0.2758695 3.196039 4.297563
#> 73  21.63854151 3.681777 0.2758695 3.196039 4.297563
#> 74  47.93985642 3.681777 0.2758695 3.196039 4.297563
#> 75  19.74103423 3.681777 0.2758695 3.196039 4.297563
#> 76  71.93558377 3.681777 0.2758695 3.196039 4.297563
#> 77   0.78847387 3.681777 0.2758695 3.196039 4.297563
#> 78  37.54899646 3.681777 0.2758695 3.196039 4.297563
#> 79  51.44077083 3.681777 0.2758695 3.196039 4.297563
#> 80   0.15705542 3.681777 0.2758695 3.196039 4.297563
#> 81  58.16040025 3.681777 0.2758695 3.196039 4.297563
#> 82  15.79052082 3.681777 0.2758695 3.196039 4.297563
#> 83  35.90283059 3.681777 0.2758695 3.196039 4.297563
#> 84  64.56318784 3.681777 0.2758695 3.196039 4.297563
#> 85  77.58233626 3.681777 0.2758695 3.196039 4.297563
#> 86  56.36468416 3.681777 0.2758695 3.196039 4.297563
#> 87  23.37033986 3.681777 0.2758695 3.196039 4.297563
#> 88   8.99805163 3.681777 0.2758695 3.196039 4.297563
#> 89   8.56120649 3.681777 0.2758695 3.196039 4.297563
#> 90  30.52183695 3.681777 0.2758695 3.196039 4.297563
#> 91  66.74265147 3.681777 0.2758695 3.196039 4.297563
#> 92   0.02388966 3.681777 0.2758695 3.196039 4.297563
#> 93  20.85699569 3.681777 0.2758695 3.196039 4.297563
#> 94  93.30341273 3.681777 0.2758695 3.196039 4.297563
#> 95  92.56447486 3.681777 0.2758695 3.196039 4.297563
#> 96  73.40943010 3.681777 0.2758695 3.196039 4.297563
#> 97  33.30719834 3.681777 0.2758695 3.196039 4.297563
#> 98  51.50633298 3.681777 0.2758695 3.196039 4.297563
#> 99  74.39746463 3.681777 0.2758695 3.196039 4.297563
#> 100 61.91592400 3.681777 0.2758695 3.196039 4.297563

# Evaluate at novel data
novel_data = data.frame(time = c(-5, 20, 300))  # Only predictors are needed
predict(demo_fit, newdata = novel_data, probs = c(0.025, 0.5, 0.975))
#>   time    predict     error       Q2.5        Q50     Q97.5
#> 1   -5   8.760546  3.829794   1.245110   8.740182  16.28175
#> 2   20   9.370080  3.794928   1.675453   9.367293  16.67768
#> 3  300 -60.126130 15.149835 -89.003556 -60.479487 -29.66752
# }