Fit Multiple Linear Segments And Their Change Points

Given a list of linear segments, mcp infers the posterior distributions of the parameters of each segment as well as the change points between segments. All segments must regress on the same x-axis. Change points are forced to be ordered using truncation of the priors. You can run fit = mcp(segments, sample=FALSE) to avoid sampling and the need for data if you just want to get the priors (fit$prior), the JAGS code fit$jags_code, or the R function to simulate data (fit$func_y).

mcp(segments, data = NULL, prior = list(), family = gaussian(),
  par_x = NULL, sample = TRUE, cores = 1, chains = 3,
  iter = 3000, adapt = 1500, update = 1500, jags_explicit = NULL,
  ...)

Arguments

segments
A list of formulas - one for each segment. The right-hand side specifices the form of intercepts and slopes. For the first segment, the left-hand side is the response variable. In the following segments, the left-hand side is the change point (on x). See examples for more details.

data
Data.frame or tibble in long format.

prior
Named list. Names are parameter names (cp_i, int_i, [x_var]_i, sigma) and the values are either

A JAGS distribution (e.g., int_1 = "dnorm(0, 1) T(0,)") indicating a conventional prior distribution. Uninformative priors based on data propertiesare used where priors are not specified. This ensures good parameter estimations, but it is a questionable for hypothesis testing. mcp uses SD (not precision) for dnorm, dt, dlogis, etc. See details. Change points are forced to be ordered through the priors using truncation, except for uniform priors where the lower bound should be greater than the previous change point, dunif(cp_1, MAXX).

A numerical value (e.g., int_1 = -2.1) indicating a fixed value.

A model parameter name (e.g., int_2 = "int_1"), indicating that this parameter is shared - typically between segments. If two varying effects are shared this way, they will need to have the same grouping variable.

family
One of gaussian(), binomial(), bernoulli(), or poission(). Only default link functions are currently supported.

par_x
String (default: NULL). Only relevant if no segments contains slope (no hint at what x is). Set this, e.g., par_x = "time".

sample
Boolean (default: TRUE). Set to FALSE if you only want to check priors, the JAGS model, etc.

cores
Positive integer or "all". Number of cores.

1: serial sampling

>1: parallel sampling on this number of cores. Ideally set chains to the same value.

"all": use all cores but one.

chains
Positive integer. Number of chains to run.

iter
Positive integer. Number of post-warmup samples to draw.

adapt
Positive integer. Number of iterations to find sampler settings.

update
Positive integer. Also sometimes called "burnin", this is the number of regular samples before anything is recorded. Use to reach convergence.

jags_explicit
Pass JAGS code to mcp to use directly. Useful if you want to make small tweaks, but mostly used for the development of mcp.

...
Further parameters for jags.fit.

Value

An mcpfit object.

Details

Noites on priors:

Order restriction is automatically applied to cp_\* parameters using truncation (e.g., T(cp_1, )) so that they are in the correct order on the x-axis UNLESS you do it yourself. The one exception is for dunif distributions where you have to do it as above.

In addition to the model parameters, MINX (minimum x-value), MAXX (maximum x-value), SDX (etc...), MINY, MAXY, and SDY are also available when you set priors. They are used to set uninformative default priors.

Use SD when you specify priors for dt, dlogis, etc. JAGS uses precision but mcp converts to precision under the hood via the sd_to_prec() function. So you will see SDs in fit$prior but precision ($1/SD^2) in fit$jags_code

See also

Examples

if (FALSE) {
# Define the segments that are separated by change points
segments = list(
  score ~ 1 + year,  # intercept + slope
   1 ~ 0 + year,  # joined slope
   1 ~ 0,  # joined plateau
   1 ~ 1  # disjoined plateau
)

# Start sampling
fit = mcp(segments, data)

# Visual inspection of the results
plot(fit)
plot(fit, "combo")

# Compare to a one-intercept-only model (no change points) with default prior
segments2 = list(1 ~ 1)
fit2 = mcp(segments2, data)  # fit another model here
fit$loo = loo(fit)
fit2$loo = loo(fit)
loo_compare(fit, fit2)

# Show all priors (not just those specified manually)
fit$prior

# Set priors and re-run
# cp_i are change points.
# int_i are intercepts.
# x_i are slopes.
# i is the segment number (change points are to the right of the segment)
prior = list(
  int_1 = "dt(10, 30) T(0, )",  # t-dist intercept. Truncated to > 0
  year_1 = "dnorm(0, 5)",  # slope of segment 1. Mean = 0, SD = 5.
  cp_2 = "dunif(cp_1, 40)",  # change point to segment 2 > cp_1.
  year_2 = "year_1",  # Shared slope between segment 2 and 1
  int_3 = 15  # Fixed intercept of segment 3
)
fit3 = mcp(segments, data, prior)

# Do stuff with the parameter estimates
fit$pars$model  # check out which parameters are inferred.
library(tidybayes)
spread_draws(fit$samples, cp_1, cp_2, int_1, year_1, year_2) %>%
   # tidybayes stuff here

# Show JAGS model
cat(fit$jags_code)
}