Extract Model Information from an mcpfit

Standard R accessors for the model formulas, family, fitting data, and number of observations stored in an mcpfit.

Usage

# S3 method for class 'mcpfit'
family(object, ...)

# S3 method for class 'mcpfit'
nobs(object, ...)

# S3 method for class 'mcpfit'
model.frame(formula, ...)

# S3 method for class 'mcpfit'
formula(x, segment = NULL, ...)

Arguments

object, x, formula

An mcpfit object.

...

Currently unused.

segment

NULL to return all segment formulas, or a positive integer selecting one segment.

Value

formula() returns the complete list of segment formulas, or one formula when segment is supplied. family() returns an mcpfamily. model.frame() returns the data retained in the fit. nobs() returns the number of fitting-data rows.

Examples

formula(demo_fit)  # Show all segment formulas
#> List of 3
#>  $ response ~ 1
#> <environment: 0x55ffb5a5c840>
#>  $ response ~ 1 ~ 0 + time
#> <environment: 0x55ffb5a5c840>
#>  $ response ~ 1 ~ 1 + time
#> <environment: 0x55ffb5a5c840>
formula(demo_fit, segment = 2)  # Show the formula for segment 2
#> response ~ 1 ~ 0 + time
#> <environment: 0x55ffb5a5c840>
family(demo_fit)  # Show the response family and link
#> 
#> Family: gaussian 
#> Link function: identity 
#> 
head(model.frame(demo_fit))  # Show the top rows of fitting data
#>    response     time
#> 1 32.842651 68.35820
#> 2 -1.160003 87.29038
#> 3 27.564248 69.01173
#> 4 10.062971 11.59361
#> 5 14.056859 19.50091
#> 6 18.292640 46.12009
nobs(demo_fit)  # Count observed response rows
#> [1] 100