Plot mcpfit

Plotting posterior fitted lines on top of data (plot(fit)) or many types of plots of parameter estimates (typically plot(fit, "combo")). See examples for typical use cases.

# S3 method for mcpfit
plot(x, type = "segments", pars = "population",
  regex_pars = character(0), facet_by = NULL, rate = TRUE,
  draws = 25, ncol = 1, ...)

Arguments

x
An mcpfit object

type
String or vector of strings. Calls bayesplot::mcmc_*type*(). Common calls are "combo", "trace", and "dens_overlay". Current options include 'acf', 'acf_bar', 'areas', 'areas_ridges', 'combo', 'dens', 'dens_chains', 'dens_overlay', 'hist', 'intervals', 'rank_hist', 'rank_overlay', 'trace', 'trace_highlight', and 'violin".

pars
Character vector. One of:

Vector of parameter names.

"population" (default) plots all population parameters.

"varying" plots all varying effects. To plot a particular varying effect, use regex_pars = "^name".

regex_pars
Vector of regular expressions. This will typically just be the beginning of the parameter name(s), i.e., "^cp_" plots all change points, "^my_varying" plots all levels of a particular varying effect, and "^cp_|^my_varying" plots both.

facet_by
String. Name of a varying group. facet_by only applies for type = "segments"

rate
Boolean. For binomial models, plot on raw data (rate = FALSE) or response divided by number of trials (rate = TRUE). If FALSE, linear interpolation on trial number is used to infer trials at a particular x. rate only applies for type = "segments"

draws
Positive integer. Number of posterior draws to use when type = "segments".

ncol
Number of columns in plot. This is useful when you have many parameters and only one plot type. ncol only when type != "segments"

...
Currently ignored.

Value

A ggplot2 object.

Details

For type = "segments", it uses fit$func_y with draws posterior samples. These represent the joint posterior distribution of parameter values.

For other type, it calls bayesplot::mcmc_*type*(). Use these directly on fit$samples if you want finer control of plotting, e.g., bayesplot::mcmc_dens(fit$samples). There are also a number of useful plots in the coda package, i.e., coda::gelman.plot(fit$samples) and coda::crosscorr.plot(fit$samples)

In any case, if you see a few erratic lines or parameter estimates, this is a sign that you may want to increase arguments 'adapt', 'update', and 'iter' in mcp.

Examples

if (FALSE) {
# Plots segments (default)
plot(fit)
plot(fit, draws = 50, rate = FALSE)  # more draws. for binomial.
plot(fit, facet_by = "my_varying")  # varying effects

# Plot parameter estimates
plot(fit, "combo")
plot(fit, "combo", pars = "varying", ncol = 3)  # plot all varying effects
plot(fit, "combo", regex_pars = "my_varying", ncol = 3)  # plot all levels of a particular varying

# More options for parameter estimates
plot(fit, "combo", pars = c("var1", "var2", "var3"), regex_pars = "^my_varying")
plot(fit, c("areas", "intervals"))

# Some plots only take pairs
plot(fit, "hex", pars = c("var1", "var2"))

# Customize one-column plots using regular ggplot2
plot(fit) + theme_bw(15) + ggtitle("Great plot!")

# Customize two-column plots using the "patchwork" package.
plot(fit, c("trace", "dens_overlay")) * theme_bw(10)
}