Summarise mcpfit objects

Summarise parameter estimates and model diagnostics.

# S3 method for mcpfit
summary(object, width = 0.95, digits = 2, prior = FALSE, ...)

fixef(object, width = 0.95, prior = FALSE, ...)

ranef(object, width = 0.95, prior = FALSE, ...)

# S3 method for mcpfit
print(x, ...)

Arguments

object
An mcpfit object.

width
Float. The width of the highest posterior density interval (between 0 and 1).

digits
a non-null value for digits specifies the minimum number of significant digits to be printed in values. The default, NULL, uses getOption("digits"). (For the interpretation for complex numbers see signif.) Non-integer values will be rounded down, and only values greater than or equal to 1 and no greater than 22 are accepted.

prior
TRUE/FALSE. Summarise prior instead of posterior?

...
Currently ignored

x
An mcpfit object.

Value

A data frame with parameter estimates and MCMC diagnostics. OBS: The change point distributions are often not unimodal and symmetric so the intervals can be deceiving Plot them using plot_pars(fit).

mean is the posterior mean

lower is the lower quantile of the highest-density interval (HDI) given in width.

upper is the upper quantile.

Rhat is the Gelman-Rubin convergence diagnostic which is often taken to be acceptable if < 1.1. It is computed using gelman.diag.

n.eff is the effective sample size computed using effectiveSize. Low effective sample sizes are also obvious as poor mixing in trace plots (see plot_pars(fit)). Read how to deal with such problems here

ts_err is the time-series error, taking autoregressive correlation into account. It is computed using spectrum0.ar.

For simulated data, the summary contains two additional columns so that it is easy to inspect whether the model can recover the parameters. Run simulation and summary multiple times to get a sense of the robustness.

sim is the value used to generate the data.

match is "OK" if sim is contained in the HDI interval (lower to upper).

Functions

fixef: Get population-level ("fixed") effects of an mcpfit object.

ranef: Get varying ("random") effects of an mcpfit object.

print.mcpfit: Print the posterior summary of an mcpfit object.

Examples

# Typical usage
summary(ex_fit)

#> Family: gaussian(link = 'identity')
#> Iterations: 3000 from 3 chains.
#> Segments:
#>   1: response ~ 1
#>   2: response ~ 1 ~ 0 + time
#>   3: response ~ 1 ~ 1 + time
#> 
#> Population-level parameters:
#>     name match  sim  mean lower  upper Rhat n.eff   ts_se
#>     cp_1    OK 30.0 31.06 23.03 38.996    1   125 317.427
#>     cp_2    OK 70.0 69.78 69.32 70.287    1  1848   0.143
#>    int_1    OK 10.0 10.32  8.93 11.778    1   429   3.400
#>    int_3    OK  0.0  0.75 -1.84  3.420    1   259  24.569
#>  sigma_1    OK  4.0  4.00  3.43  4.589    1  1056   0.160
#>   time_2    OK  0.5  0.54  0.40  0.683    1   140   0.091
#>   time_3    OK -0.2 -0.23 -0.39 -0.063    1   255   0.082

summary(ex_fit, width = 0.8, digits = 4)  # Set HDI width

#> Family: gaussian(link = 'identity')
#> Iterations: 3000 from 3 chains.
#> Segments:
#>   1: response ~ 1
#>   2: response ~ 1 ~ 0 + time
#>   3: response ~ 1 ~ 1 + time
#> 
#> Population-level parameters:
#>     name match  sim    mean   lower   upper   Rhat n.eff     ts_se
#>     cp_1    OK 30.0 31.0598 26.2296 35.8264 1.0030   125 317.42685
#>     cp_2    OK 70.0 69.7816 69.4614 70.2747 1.0000  1848   0.14288
#>    int_1    OK 10.0 10.3207  9.5033 11.3206 1.0007   429   3.40038
#>    int_3    OK  0.0  0.7475 -1.0863  2.3992 1.0007   259  24.56869
#>  sigma_1    OK  4.0  4.0042  3.6026  4.3419 0.9996  1056   0.15986
#>   time_2    OK  0.5  0.5442  0.4611  0.6251 1.0024   140   0.09139
#>   time_3    OK -0.2 -0.2327 -0.3474 -0.1296 1.0000   255   0.08183

# Get the results as a data frame
results = summary(ex_fit)

#> Family: gaussian(link = 'identity')
#> Iterations: 3000 from 3 chains.
#> Segments:
#>   1: response ~ 1
#>   2: response ~ 1 ~ 0 + time
#>   3: response ~ 1 ~ 1 + time
#> 
#> Population-level parameters:
#>     name match  sim  mean lower  upper Rhat n.eff   ts_se
#>     cp_1    OK 30.0 31.06 23.03 38.996    1   125 317.427
#>     cp_2    OK 70.0 69.78 69.32 70.287    1  1848   0.143
#>    int_1    OK 10.0 10.32  8.93 11.778    1   429   3.400
#>    int_3    OK  0.0  0.75 -1.84  3.420    1   259  24.569
#>  sigma_1    OK  4.0  4.00  3.43  4.589    1  1056   0.160
#>   time_2    OK  0.5  0.54  0.40  0.683    1   140   0.091
#>   time_3    OK -0.2 -0.23 -0.39 -0.063    1   255   0.082

# Varying (random) effects
# ranef(my_fit)

# Summarise prior
summary(ex_fit, prior = TRUE)

#> Family: gaussian(link = 'identity')
#> Iterations: 3000 from 3 chains.
#> Segments:
#>   1: response ~ 1
#>   2: response ~ 1 ~ 0 + time
#>   3: response ~ 1 ~ 1 + time
#> 
#> Population-level parameters:
#>     name match  sim    mean    lower  upper Rhat n.eff   ts_se
#>     cp_1    OK 30.0 50.6712  4.5e+00  93.54    1  3434 6.7e+02
#>     cp_2    OK 70.0 74.3771  3.2e+01  97.96    1  3461 4.5e+02
#>    int_1    OK 10.0  0.2158 -1.2e+02 102.31    1  2699 3.4e+03
#>    int_3    OK  0.0 -1.2566 -1.0e+02 105.00    1  3026 2.7e+03
#>  sigma_1    OK  4.0  9.8986  7.3e-03  23.52    1  3000 4.8e+01
#>   time_2        0.5  0.0023 -4.5e-01   0.39    1  3326 4.5e-02
#>   time_3    OK -0.2  0.0057 -3.7e-01   0.45    1  3000 5.1e-02